# api-admin/src/main/resources/api/openapi.yml
openapi: 3.0.3
info:
  title: Juniper Admin API
  description: API for portal and study management
  version: 0.0.1
paths:
...
  /api/portals/v1/{portalShortcode}/adminUser/{adminUserId}:
    delete:
      summary: Remove an admin user for portal
      tags: [ portal ]
      operationId: removePortalUser
      parameters:
        - { name: portalShortcode, in: path, required: true, schema: { type: string } }
        - { name: adminUserId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204':
          description: admin user removed
        '404':
          description: admin user not found
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  responses:
...
    ServerError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorReport'
  schemas:
...
    ErrorReport:
      type: object
      required: [ message, statusCode ]
      properties:
        message:
          type: string
        errorClass:
          type: string
        statusCode:
          type: integer
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security:
  - bearerAuth: [ ]